+++ service/admin/sensorDataApi.ts
